refactor: rename query
