+++ apps/main/src/lend/components/ChartAndActivityComp/index.tsx
@@ -1,4 +1,3 @@
-import { useOneWayMarket } from '@/lend/entities/chain'
 import { useOhlcChartState } from '@/lend/hooks/useOhlcChartState'
 import { networks } from '@/lend/networks'
 import { Api, ChainId } from '@/lend/types/lend.types'
@@ -10,6 +9,7 @@ import type { Address, Token } from '@primitives/address.utils'
 import type { Decimal } from '@primitives/decimal.utils'
 import { useBandsChartVisible } from '@ui-kit/hooks/useLocalStorage'
 import type { Range } from '@ui-kit/types/util'
+import { useLendMarket } from '../../hooks/useLendMarket'
 
 type ChartAndActivityCompProps = {
   rChainId: ChainId
@@ -20,7 +20,7 @@ type ChartAndActivityCompProps = {
 
 export const ChartAndActivityComp = ({ rChainId, marketId, api, previewPrices }: ChartAndActivityCompProps) => {
   const [isBandsVisible] = useBandsChartVisible()
-  const market = useOneWayMarket(rChainId, marketId).data
+  const market = useLendMarket(rChainId, marketId).data
   const collateralTokenAddress = market?.collateral_token.address
   const borrowedTokenAddress = market?.borrowed_token.address
