--- apps/main/src/lend/entities/chain/index.ts
@@ -1,2 +1,0 @@
-export { useChainId } from './chain-info'
-export { useOneWayMarket } from './chain-hooks'
