+++ apps/main/src/lend/hooks/useSupplyTotalApr.ts
@@ -1,14 +1,14 @@
 import { useMemo } from 'react'
 import { zeroAddress } from 'viem'
-import { useOneWayMarket } from '@/lend/entities/chain'
 import { useStore } from '@/lend/store/useStore'
 import { ChainId, MarketRates, RewardOther, MarketRewards } from '@/lend/types/lend.types'
 import { getTotalApr } from '@/lend/utils/utilsRewards'
 import { useMarketRates } from '@/llamalend/queries/market'
 import { FORMAT_OPTIONS, formatNumber } from '@ui/utils'
+import { useLendMarket } from '../hooks/useLendMarket'
 
 export function useSupplyTotalApr(rChainId: ChainId, marketId: string) {
-  const market = useOneWayMarket(rChainId, marketId).data
+  const market = useLendMarket(rChainId, marketId).data
   const marketRewardsResp = useStore(state => state.markets.rewardsMapper[rChainId]?.[marketId])
   const marketRatesResp = useStore(state => state.markets.ratesMapper[rChainId]?.[marketId])
   const {
