+++ apps/main/src/lend/queries/invalidateMarketDetails.ts
