+++ apps/main/src/lend/queries/lend-market-names.query.ts
@@ -1,16 +1,16 @@
+import { requireLib } from 'curve-ui-kit/src/features/connect-wallet'
+import { ChainParams, type ChainQuery, queryFactory, rootKeys } from 'curve-ui-kit/src/lib/model/query'
+import { llamaApiValidationSuite } from 'curve-ui-kit/src/lib/model/query/curve-api-validation'
+import { Chain } from 'curve-ui-kit/src/utils'
 import { USE_API } from '@/llamalend/queries/market/market.constants'
-import { notFalsy } from '@primitives/objects.utils'
-import { requireLib } from '@ui-kit/features/connect-wallet'
-import { ChainParams, type ChainQuery, queryFactory, rootKeys } from '@ui-kit/lib/model/query'
-import { llamaApiValidationSuite } from '@ui-kit/lib/model/query/curve-api-validation'
-import { Chain } from '@ui-kit/utils'
+import { notFalsy } from '@curvefi/primitives/objects.utils'
 
 const V1 = 'v1' as const
 const V2 = 'v2' as const
 
 const v2chains = [Chain.Optimism]
 
-export const { useQuery: useOneWayMarketNames, prefetchQuery: prefetchMarkets } = queryFactory({
+export const { useQuery: useLendMarketNames, prefetchQuery: prefetchMarkets } = queryFactory({
   queryKey: ({ chainId, enableLLv2 }: ChainParams & { enableLLv2: boolean }) =>
     [...rootKeys.chain({ chainId }), 'lendMarkets.getMarketList', { enableLLv2 }] as const,
   queryFn: async ({ chainId, enableLLv2 }: ChainQuery & { enableLLv2: boolean }): Promise<string[]> => {
