+++ apps/main/src/lend/store/createAppSlice.ts
@@ -2,13 +2,13 @@ import { produce } from 'immer'
 import lodash from 'lodash'
 import type { Config } from 'wagmi'
 import { StoreApi } from 'zustand'
-import { prefetchMarkets } from '@/lend/entities/chain/chain-query'
 import type { State } from '@/lend/store/useStore'
 import { Api, Wallet } from '@/lend/types/lend.types'
 import { recordEntries } from '@primitives/objects.utils'
 import { isLLv2Enabled } from '@ui-kit/hooks/useFeatureFlags'
 import { log } from '@ui-kit/lib/logging'
 import { ReleaseChannel } from '@ui-kit/utils'
+import { prefetchMarkets } from '../queries/lend-market-names.query'
 
 export type SliceKey = keyof State | ''
 export type StateKey = string
