+++ tests/cypress/component/llamalend/manage-soft-liq.rpc.cy.tsx
@@ -1,5 +1,5 @@
 import { type ReactNode, useMemo } from 'react'
-import { prefetchMarkets } from '@/lend/entities/chain/chain-query'
+import { prefetchMarkets } from '@/lend/queries/lend-market-names.query'
 import { ClosePositionForm } from '@/llamalend/features/manage-soft-liquidation/ui/tabs/ClosePositionForm'
 import { ImproveHealthForm } from '@/llamalend/features/manage-soft-liquidation/ui/tabs/ImproveHealthForm'
 import type { NetworkDict } from '@/llamalend/llamalend.types'
